+++ src/replique/repl_cljs.clj
@@ -35,8 +35,8 @@
        cljs-qualifier :qualifier} cljs.util/*clojurescript-version*]
   (assert (or (> cljs-major 1) (> cljs-minor 9)
               (and (= cljs-major 1) (= cljs-minor 9)
-                   (>= cljs-qualifier 183)))
-          (format "Replique is compatible with clojurescript 1.9.183+, current version is: %s.%s.%s" cljs-major cljs-minor cljs-qualifier)))
+                   (>= cljs-qualifier 473)))
+          (format "Replique is compatible with clojurescript 1.9.473+, current version is: %s.%s.%s" cljs-major cljs-minor cljs-qualifier)))
 
 (declare init-repl-env)
 (declare init-compiler-env)