+++ CHANGES.md
@@ -1,12 +1,13 @@
 # Version 0.0.8
 
+- Fix a "broken link" exception when reloading a web browser page connected to the browser REPL
 - The omniscient no longer always use the namespace of the session when resolving symbols 
 - Replique now checks the clojurescript version number for compatibility
 - The browser REPL now supports query string parameters
 - Stringify all Clojure symbols in tooling messages
 - Starting a REPL did not work when some dependencies were excluded
 - Add the remove-var interactive macro
-- Update the minimum cljs version required to 1.9.183
+- Update the minimum cljs version required to 1.9.473
 - Improve local bindings analysis
 - Fix the omniscient debugger when updating the var being debugged in an already started omniscient repl